Delete

+++ src/main/java/com/luv2code/springboot/cruddemo/service/EmployeeServiceImpl.java
@@ -51,6 +51,10 @@ public class EmployeeServiceImpl implements EmployeeService {
     @Override
     @Transactional
     public void delete(int id) {
+        Employee employeeToDelete = employeeDAO.findById(id);
+        if (employeeToDelete == null) {
+            throw new EmployeeNotFoundException("Employee not found");
+        }
         employeeDAO.delete(id);
     }
